dpdk_module: Fix check link status function
In previous implementation, link status was being checked sequentially
for the number of ports initialized. This was working if we were not
skipping any port in between. But, If we skiped one or more ports,
link status was being checked for the ports which were not initialized
and skipping the ones which were initialized.

This issue have been fixed by using the correct port ids of the
initialized ports.

Signed-off-by: Rohit Raj <rohit.raj@nxp.com>
DPDK-4043